flask stego project/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Pillow==10.4.0
pycryptodome==3.20.0
